refactor: adiciona verificação de registro salvo no banco

+++ src/main/java/me/gabriel/webflux/application/config/DatabaseInitializer.java
@@ -21,13 +21,17 @@ public class DatabaseInitializer implements CommandLineRunner {
 
   @Override public void run(String... args) {
 
-    playlistRepository.deleteAll()
-      .thenMany(
-        Flux.just("API REST Spring Boot", "Deploy de uma aplicação java no IBM Cloud", "Java 8",
-                  "Github", "Spring Security", "Web Service RESTFULL", "Bean no Spring Framework"
-          )
-          .map(name -> new PlaylistDocument(UUID.randomUUID().toString(), name))
-          .flatMap(playlistRepository::save)
-      ).subscribe(System.out::println);
+    var notHavePlaylistRegistered = playlistRepository.count().blockOptional().orElse(0L) == 0;
+
+    if(notHavePlaylistRegistered) {
+      playlistRepository.deleteAll()
+        .thenMany(
+          Flux.just("API REST Spring Boot", "Deploy de uma aplicação java no IBM Cloud", "Java 8",
+                    "Github", "Spring Security", "Web Service RESTFULL", "Bean no Spring Framework"
+            )
+            .map(name -> new PlaylistDocument(UUID.randomUUID().toString(), name))
+            .flatMap(playlistRepository::save)
+        ).subscribe(System.out::println);
+    }
   }
 }
